requirements.txt
flask
requests
jieba
numpy
scipy
matplotlib
pywebview==3.6.3
pyinstaller==4.5.1
